TOOLS-1999 Respect --ssl if not seen in --uri
This commit fixes a bug where the use of --uri overrode the use of
the --ssl command line option.  Now, --ssl is respected unless "ssl=..."
explicitly appears in the URI options.  If "ssl=false" and --ssl is
used, a conflicting arguments error occurs.

Also fixes the test.sh script to run tests on common/*.