minor: rationalize Evergreen QA tests with tags
This commit moves tests that weren't being run because they require
multiple versions to a separate directory so that it's clearer they
are being skipped.  The remaining QA tests that require "version X or
later" have their tags updated like 'requires_min_mongo_42'.
Evergreen task exclusions are updated accordintly.

The Evergreen config renames cert testing and the "with archive" and
"with gzip" variants to make it clearer they only need to test a single
version and -- in the case of archive/gzip -- that they are variants of
the normal dump/restore tests.