TOOLS-2268 Add --mode=delete option
When mode is set to delete, mongoimport will delete existing documents
in the database with matching documents from the import file.
No operation will be performed with non-matching documents. By default,
the _id field will be used to match documents. Use --upsertFields to
specify the fields against which to match existing documents.