TOOLS-1954 Add --useArrayIndexFields flag (#175)
When --useArrayIndexFields is set, natural numbers in fields are interpreted as array indexes. For example:

a.0,a.1
1,2

would result in the document { "a": [ 1, 2 ] }. Indexes must increase sequentially and must start from 0. For example, fields a.1,a.0 would fail. Fields a.0,a.2 would also fail since an element is skipped. However, a.0.b,a.1.c,a.0.d actually will not fail (although this type of out-of-order fields should be discouraged).

A blank value for a field is inserted into the array as an empty string. E.g.

a.0,a.1
,2

Would result in the document: { "a": [ "", 2 ] }.

Numerical keys with leading 0s are not interpreted as array indexes. E.g.

a.000,a.001
1,2

would result in { "a": { "000": 1, "001": 2 } }. 

If the first part of a key is a natural number, it is interpreted as a document key since the top level of a BSON document is (of course) a document. E.g.

0,1
1,2

would result in { "0": 1, "1": 2 }.

Field names are incompatible if they imply different types of value at a certain key. E.g. the following would be incompatible:

- a.0 (a is an array) and a.a (a is a document)
- a (a is a value) and a.0 (a is an array). 

--useArrayIndexFields only has an effect for CSV and TSV file types since arrays are already encoded in JSON.