TOOLS-2958 Fix bsonutil.IsEqual (#427)
This commit includes the following changes:

## Fix `bsonutil.IsEqual`

A previous change broke this function so it always returned true. This was due to a bug that had it always compare the left value to itself instead of to the right value.

I also added tests for this function and converted the existing tests in `bsonutil_test.go` to use testify.

## Don't allow `DeleteIndexes` to delete >1 index when matching by key

A key should only ever have 0 or 1 matches, but there was a bug in `bsonutil.IsEqual` that caused us to match all keys. Adding this extra check helps prevent index dropping disasters in the future.

## Add some tests for `IndexCatalog.DeleteIndexes`

This also fixes a bug with a command to drop all indexes:

```
{{"dropIndexes": "collection"}, {"index": "*"}}
```

This would drop the index on the `_id` field if that index was not named `_id_`, which is a possibility with clustered collections.