Fix client certificate chain bug
If the user provides a client certificate chain (i.e., a client cert + an
intermediate cert), currently mongodump will ignore all certs except the last
one (i.e., dropping the client cert) and then compare the intermediate cert to
the private key, resulting in a fatal error about the key and the cert not
matching. Additionally, since we drop all but the last cert, we don't present
the server with the correct client cert.

Here we append all certs in the cert/key file together + use the first cert for
the subject name, which is standard practice. This fixes X.509 auth for
mongodump.