Move prompting for password to earlier in options processing
Previously, we didn't prompt for a password until we attempted to establish a
database connection. But before we did that we would call
`connstring.ConnString.Validate`, which would return an error if the user for
the connection was set but no password was provided.

This commit moves the prompt for the password to just before the call to
`Validate`, so we should never end up complaining about a missing password
before having a chance to prompt the user.

I tested this by hand with a few different patterns:

$> go run ./mongotop/main/ -u user --authenticationMechanism 'PLAIN'
$> go run ./mongotop/main/ -u user --password="" --authenticationMechanism 'PLAIN'
$> go run ./mongotop/main/ -u user --password=""

I tried this with a few different tools, including mongodump, mongorestore,
and mongotop.