Add a tool to generate github_pr_aliases for the Evergreen config
The aliases define the default tasks per variant that are run for all PRs by
default. The rules by which we select these are somewhat complex, so
maintaining this by hand would be painful. This code will look at the existing
config and generate this based on the rules. This will make it easier to keep
these aliases up to date as we modify our tasks and variants.