TOOLS-3185 Fix `TestMongorestoreTxns` failure with Server 6.1+ (#474)
The old hard-coded data no longer worked with Server 6.1 as the format for
updates has changed.

This PR adds a new file with the new format to be used with Server 6.1+.

In the future, we might want to instead generate this data in the test. That
was the first approach I tried, but it's not possible to generate many
transaction scenarios with the Server API as it stands. Specifically,
generating a set of partial transactions or `abortTransaction` and
`commmitTransaction` ops isn't possible, as these are only created based on
Server internals.