Remove test:srv test type
There was exactly one test of this type. I made this a unit test instead. If
the `ATLAS_URI` is not set, the test will be skipped.

Running both `test:integration` and `test:srv` in Evergreen produced confusing
logs, since this tries to execute _every_ test twice. But the second time,
with `test:srv`, all the tests but one are skipped. This confuses Evergreen,
which does not handle seeing the same test name twice.