TOOLS-3203: Revamp config-exclusion logic. (#543)
mongodump & mongorestore currently have logic for determining which
collections in the `config` database to include or exclude. This changeset
rewrites that logic as an allow-list of specific collection names in
order to avoid backing up or restoring data that doesn’t really belong
in backup/restore workflows. (See the JIRA ticket for details.)

This also slightly reorders the dump-intents gathering logic to preserve
detection of resharding.