TOOLS-3365 Prep release for 100.7.5 (#596)
Adds a new build target checkMinVersion which checks the version of Go being used. It will error if it is not at least the minimum version set in build.go. checkMinVersion is run during the dist task before binaries are built.

Removes support for PPC Ubuntu 18.04 since it is no longer supported as of 4.2.17 and 4.4.10. Evergreen builds are no longer run for this distro.

Changed the SUSE 15 build host to suse15sp4 since this is the recommended distro and has go 1.19.12.

Reverts #582 since a bug in goconvey was causing test failures.