TOOLS-3411 Fix issue where sort order of views was not always preserved on dump (#618)
Fixes an issue where mongodump did not always preserve the order of options for a view or collection. This could result in different sort orders than expected on views. It is easily fixed by moving from the unordered bson.M type to the ordered bson.D type for collection options. Index options are still a bson.M, but the keys are bson.D so they are not affected by this issue.