Remove use of `filemd5` command in test code
This seems to have been related to something the `mongofiles` code _used_ to
do quite some time in the past. There are no longer any references to the
`filemd5` command anywhere else in the tools codebase.

The `filemd5` command if going to be deprecated and eventually removed from
the Server.