TOOLS-3500 Remove use of the `filemd5` command in test code (#640)
This command is deprecated and will be removed in a future Server release. We were using that command to test that a file uploaded and downloaded via {{mongofiles}} is unchanged. This has been replaced by simply checking this entirely in the test's JS code.

This also updates the `github_pr_aliases` in the Evergreen config, which had gotten out of date.