TOOLS-3533 Update most Go packages and update Go to 1.21.10 (#647)
This updates Go to 1.21.10 and most of our Go packages. I've held back the `golint` fork used for linting. The "new" (still 10 year old) version complains about a bunch of stuff we don't want to change. We will probably get rid of this particular linter entirely in a future PR.