Fix lint-js and SRV test failures and clean up options testing code (#651)
This PR fixes an issue with the Evergreen config for our lint-js testing.

It fixes the SRV tests by mocking DNS so we no longer rely on a particular cluster existing in Atlas when the tests are run. Since these tests are no longer special, the `test:srv` build target is gone and these tests are just run as normal unit tests.

It also cleans up the options testing code to use testify.