TOOLS-3535 Add gosec as a linter and generate a SARIF report as part of release (#655)
This PR adds infrastructure to install dev tools including `precious`, which we use to run `gosec` as a linter. As part of the release process, this will generate a SARIF report. If there are new issues in the code base, then the release will be blocked.

It also updates the `lint-go` Evergreen task to run our new linting, which right now will just run `gosec`.

This also includes comments to disable gosec's linting in a few spots. These comments end up in the eventual SARIF report explaining why we are suppressing a particular error.