TOOLS-3550 Add the augmented SBOM file for the 100.9.5 release and add a release check for this file (#657)
This adds the augmented SBOM for the next release.

It also updates our Evergreen config so that it checks that this file exists as part of the release. If it doesn't, the release will be aborted.

Finally, it adds a new Evergreen parameter, `fake_tag_for_release_testing`, which can be used to test things like this new task. It will set the `EVG_TRIGGERED_BY_TAG` env var without actually triggering a release, which an actual tag would do.