TOOLS-3559 Add the Augmented to all release artifacts (tarballs, debs, etc.) (#674)
This also significantly changes the release code so that the files generated by the `dist` task following the naming scheme of the file we would/will release. For non-tag builds, these files will include the git commit in their name. When we do a release, the files that we release and make available for download should only include the tag.