Fix the `native-cert-ssl-4.4` task (#659)
This updates the `ssl_with_system_ca.js` JS test to match the code from the Server's 4.4.29 tag. This fixes issues with how the test was configuring the Server. Also, the test was simply not using the client certs that it needs to use in order to make this test actually test anything.