TOOLS-3572 Fix the version used for Debian packages and add tests for version code (#687)
For non-stable releases, we want to use `VERSION~DATE.COMMIT` (e.g. ` mongodb-database-tools_100.9.5+20240619.d8f2d106_amd64.deb`) in the Debian package name. For stable releases, the commit should not be included.

This also removes the `Pre` field from the version. This was no longer used with our current tagging scheme, because we're not doing rc releases.