Add a new `lint` task to replace the `lint-go` task and use `golangci-lint` (#679)
This will eventually replace the existing `lint-go` and `lint-js` tasks entirely.

This replaces our ancient `golint` with `golangci-lint` and fixes a _lot_ of linter issues, mostly:

* Unused functions, vars, and arguments
* Missing error checking (I added a lot of `//nolint` comments where it was too painful to add actual checking)
* Passing `nil` as a context
* Using the deprecated `ioutil` package

This also removes the `format-go` task and uses `precious` to check formatting instead.

We ignore all deprecation issues reported by `staticcheck` for now. There are a huge number of these because of the Go driver upgrade. There is a separate ticket, TOOLS-3633, for fixing these.

---------

Co-authored-by: Jian Guan <61915096+tdq45gj@users.noreply.github.com>