Handle 'system.buckets' prefix correctly when .bson is target
When the bson file is passed as the target, CreateIntentForCollection
does not set the collection name correctly by removing the
"system.buckets." prefix. This will cause errors like:
"Invalid namespace: test.system.buckets.system.buckets" as a result of
adding the same prefix twice to the namespace.

Making this change in the processing of the file so that the intent can
be initialized correctly rather than remove the prefix in the intent, so
that all code that intentionally accesses the collection name without
the prefix (when dropping the collection for example) works correctly.