TOOLS-3719: fail if --restoreDbUsersAndRoles used on wrong target (#758)
If the target was dumped with --dumpDbUsersAndRoles then we would find
the db specific users / roles and the intents would not be nil, so we
can use that check after the intents have been created to make sure that
this option is used correctly.

Adding integration tests to mongorestore_test.go, to replace
test/qa-tests/jstests/restore/empty_users_and_roles.js and add more test
cases.

---------

Co-authored-by: Felipe Gasper <FGasper@users.noreply.github.com>