TOOLS-3540 Fix handling of empty timestamps (#767)
Because the server generally interprets Timestamp(0, 0) to mean “current
time”, previously mongorestore has errantly been restoring backups with
this value incorrectly.

The server has added an experimental mechanism to preserve such values
easily; however, that is only available in recent builds. For 4.4 and
4.2, though—which mongorestore still supports—we need a backup strategy.

It happens that aggregation pipelines will, in certain configurations,
apply zero timestamps literally. This changeset uses that in order to
achieve correct restorations of zero timestamps in 4.4 and 4.2: first it
inserts the document, then it updates it with a $replaceRoot
aggregation.