TOOLS-3782: handle missing authSchema doc in 8.1 (#771)
In SERVER-83663, the authSchema doc in system.version was removed
because server 8.1 only supports one auth version (5) and the doc is not
required anymore. mongodump and restore expect the authSchema document
to exist in the server in order to use the auth version and determine
the compatibility of auth versions between the source and destination
servers.

Since there's only one authVersion in 8.1+, we set it to 5 based on the
server version. In mongodump's VerifySystemAuthVersion, in place of
checking the authSchema doc we check whether users and roles exist in
the server, because that was the condition for an authSchema doc to have
been written in older server versions.

All server version checks modified or added in this change will already
work as expected for archive mode.
Unfortunately, we don't currently have a way to find the dump's source
server version in mongorestore in bson / directory mode. mongodump does
not write the server's version into the dump unless it's used in
--archive mode. As a result, we're making mongodump write the server
version and tool version into a new prelude.json file in the dump
directory, and mongorestore can read from this and use it to determine
the dump's server version. If the dump's server version is 8.1.0+, we
default to authSchema 5 in GetDumpAuthVersion.

The new prelude.json is intended as the equivalent of the Prelude struct
that we currently write into the archive. For now we're just adding the
data that we need, but in the future we can expand it to add other
metadata about the dump.