TOOLS-3764: Run integration tests against server version 8.1 (#770)
This commit contains the changes required to run the mongo tools tests
against mongo server version 8.1.

Changes to JS test builds:
After SERVER-81339, ShardingTest, and ReplSetTest are not shipped /
loaded with the mongo shell by default. They are now modules that need
to be imported into the shell environment. We're copying over the files
`replsettest.js`, `shardingtest.js` and `parallelTester.js` from the
mongo repo into the `test/shell_common/libs/` directory. Small changes
were required in both `ReplSetTest` and `ShardingTest` so that they can
be used in the context of our JS tests. The changes to the original
versions are listed at the top of each file.

We're loading the libs in `test/shell_common/` into the mongo shell for
tests by calling `load_libs.8.1.js` in an `--eval` statement, and the
evergreen builds and JS tests have been set up to look for a version
specific `load_libs.X.js` file depending on server / shell version.
`load_libs.8.1.js` expects the shell to be run from `test/qa-tests` or
`test/legacy42`, so we're adding a `test/shell_common/jsconfig.json`
that is hidden away during jstests, but moved into the base directory to
run `mongo` from there during `create repl_set`. This is temporary, as
we should just remove the use of `ReplSetTest` to create replica sets.

Additionally - `replsettest.8.0.js` was added because we thought it was
required in the 8.0.5 shell, but tests are passing without it so we're
removing it.

Notable changes to tests to work with 8.1.0 mongo server / shell:
- SERVER-83663 - The authSchema was removed in server 8.1, so we have to
skip checking for it in `dumprestore_auth2.js`
- SERVER-100146 - We were calling `assert.eq` with more than a bunch of
extra string arguments, expecting that they will all be joined into one,
but instead the shell was just ignoring them. An unrelated change in 8.1
shell adding another 4th argument now causes assert calls to fail if
they have too many arguments, so we're converting all the strings into a
single string.
- SERVER-88158 - `applyOps` does not upsert by default, and therefore we
need to insert the document into the collection first before trying to
update it in `preserve_oplog_structure_order.js`
- SERVER-96103 - `mongofiles_get.js`, Windows specific ecaping the
quotes in the get_id JSON input is not required for 8.1.0
- SERVER-95434 - `shardOptions` has been removed, so we use `rsOptions`
instead in `topology_helper.js`. This should be a no-op for older shell
versions that use `shardOptions`.