Use // operator correctly to set empty when generating sbom
The left side of the alternative operator (//) in jq will not be empty
in this case because of the string being appended. This causes
"pkg:golang/github.com/mongodb/mongo-tools@" to get included in the list
of dependencies for mongo-tools, without the version after "@", even
though the `go list -json` output for mongo-tools does not contain a
.Module.Version field for mongo-tools.

Compared against the old file this script generates to sends to silkbomb
and the only thing removed was
"pkg:golang/github.com/mongodb/mongo-tools@"

Including the new sbom and augmented sbom in this commit.