TOOLS-3941 - make mongodump_passthrough work with newer pinned mongosync (#830)
The devprod tiger team effort made various good changes to mongosync
evergreen config and scripts.

This commit adjusts the mongo-tools/mongodump_passthrough scripts to
work against a pinned newer mongosync:

- shell scripts have been moved from src/mongosync/evergreen/foo.sh to
src/mongosync/evergreen/scripts/foo.sh
  
- mongodump_passthrough tests now run with build variant
"mongodump_passthru_v", to isolate the specifics of the
hardware platform from the name used in scripts and configuration of
Build Baron context rules

- "mongodump_passthru_v" is a cheaper amazon2-arm64 instance type, as
used in mongosync to reduce costs.
  
- mongodump_passthrough jstestfuzz scripts are currently disabled.
There's some issue with the new world that I
don't have time to solve this week, and I'm on PTO next week, so this PR
solves 3/4 of the problem and I'll file a
    follow-up ticket to fix jstestfuzz later.