TOOLS-3964: omit the main module from go list for sbom generation (#841)
We fixed this earlier by omitting the `.Module` structs that didn't
contain `.Path`, but that was removed by mistake in #810.
Rather than add it back and depend on `.Path` being empty, I'm adding a
filter to remove the main module. This is indicated in the go list json
by `.Main` being set to true in the `.Module` struct.