TOOLS-4019 Add safeguards around mongofiles’s file restoration (#853)
This adds 2 important safeguards in mongofiles:
- When `get`ting files by either a regular expression or ID, and the
filename resolves to a path outside the current directory, mongofiles
now refuses to restore the files. This prevents inadvertent creation of
files in unintended locations. A new `--allowUnsafeTraversal` flag
allows this to be overridden. (NB: mongofiles still always restores
files that are _under_ the current directory.)
- mongofiles no longer overwrites existing files by default. If a file
already exists, mongofiles will fail the restoration. This can be
overridden with a new `--overwriteLocal` flag.

*IMPORTANT:* This constitutes a breaking change to the CLI.