Compute minimum Go version dynamically; set it to 1.24. (#854)
This replaces a hard-coded (and easily overlooked!) minimum Go version
with just parsing the `go.mod` file.

It also reduces the minimum from 1.25 to 1.24. This works around an
as-yet-unsolved build issue, which unblocks the next release.

Also in the interest of unblocking the release: this makes the
`mongofiles_get.js` test accommodate 8.0 & 7.0 versions that received
SERVER-96103’s backport.